+++ R/runs.R
@@ -1,9 +1,17 @@
 #' Runs
 #'
 #' @description
-#' `vec_identify_runs()` returns a vector of identifiers for the elements of
-#' `x` that indicate which run of repeated values they fall in. The number of
-#' runs is also returned as an attribute, `n`.
+#' - `vec_identify_runs()` returns a vector of identifiers for the elements of
+#'   `x` that indicate which run of repeated values they fall in. The number of
+#'   runs is also returned as an attribute, `n`.
+#'
+#' - `vec_run_sizes()` returns an integer vector corresponding to the size of
+#'   each run. This is identical to the `times` column from `vec_unrep()`, but
+#'   is faster if you don't need the run keys.
+#'
+#' - [vec_unrep()] is a generalized [base::rle()]. It is documented alongside
+#'   the "repeat" functions of [vec_rep()] and [vec_rep_each()]; look there for
+#'   more information.
 #'
 #' @details
 #' Unlike [base::rle()], adjacent missing values are considered identical when
@@ -13,14 +21,22 @@
 #' @param x A vector.
 #'
 #' @return
-#' An integer vector with the same size as `x`. A scalar integer attribute,
-#' `n`, is attached.
+#' - For `vec_identify_runs()`, an integer vector with the same size as `x`. A
+#'   scalar integer attribute, `n`, is attached.
 #'
-#' @export
+#' - For `vec_run_sizes()`, an integer vector with size equal to the number of
+#'   runs in `x`.
+#'
+#' @seealso
+#' [vec_unrep()] for a generalized [base::rle()].
+#'
+#' @name runs
 #' @examples
 #' x <- c("a", "z", "z", "c", "a", "a")
 #'
 #' vec_identify_runs(x)
+#' vec_run_sizes(x)
+#' vec_unrep(x)
 #'
 #' y <- c(1, 1, 1, 2, 2, 3)
 #'
@@ -31,14 +47,26 @@
 #' )
 #'
 #' vec_identify_runs(df)
+#' vec_run_sizes(df)
+#' vec_unrep(df)
+NULL
+
+#' @rdname runs
+#' @export
 vec_identify_runs <- function(x) {
-  .Call(ffi_vec_identify_runs, x)
+  .Call(ffi_vec_identify_runs, x, environment())
+}
+
+#' @rdname runs
+#' @export
+vec_run_sizes <- function(x) {
+  .Call(ffi_vec_run_sizes, x, environment())
 }
 
 vec_locate_run_bounds <- function(x, start = TRUE) {
-  .Call(ffi_vec_locate_run_bounds, x, start)
+  .Call(ffi_vec_locate_run_bounds, x, start, environment())
 }
 
 vec_detect_run_bounds <- function(x, start = TRUE) {
-  .Call(ffi_vec_detect_run_bounds, x, start)
+  .Call(ffi_vec_detect_run_bounds, x, start, environment())
 }