+++ src/decl/rep-decl.h
@@ -54,7 +54,4 @@ static inline
 void stop_rep_each_times_oob(int times, r_ssize i, struct r_lazy call, struct vctrs_arg* p_times_arg);
 
 static
-r_obj* vec_unrep(r_obj* x);
-
-static
-r_obj* new_unrep_data_frame(r_obj* key, r_obj* times, r_ssize size);
+r_obj* vec_unrep(r_obj* x, struct r_lazy error_call);