+++ NEWS.md
@@ -1,5 +1,9 @@
 # vctrs (development version)
 
+* New `vec_run_sizes()` for computing the size of each run within a vector. It
+  is identical to the `times` column from `vec_unrep()`, but is faster if you
+  don't need the run key (#1210).
+
 # vctrs 0.5.2
 
 * New `vec_expand_grid()`, which is a lower level helper that is similar to